requirements/readthedocs.txt
# Before installing mmcv-full, make sure that PyTorch has been successfully installed following the official guide.
# pip install mmcv-full==1.3.0 -f https://download.openmmlab.com/mmcv/dist/cu102/torch1.6.0/index.html
mmcv-full==1.3.0 
torch
torchvision
